requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
numpy==2.2.6
opencv-contrib-python==4.12.0.88
paho-mqtt==2.1.0
requests==2.32.5
urllib3==2.5.0
